Migrate src/lib/oncstants/index.js
