+++ packages/emoji-picker-react/src/lib/constants/index.ts
@@ -1,0 +1,28 @@
+export const EMOJI_PROPERTY_NAME = 'n';
+export const EMOJI_PROPERTY_UNIFIED = 'u';
+export const EMOJI_PROPERTY_SKIN_VARIATIONS = 'v';
+export const EMOJI_PROPERTY_GROUP = 'g';
+
+export const GROUP_NAME_PEOPLE = 'smileys_people';
+export const GROUP_NAME_NATURE = 'animals_nature';
+export const GROUP_NAME_FOOD = 'food_drink';
+export const GROUP_NAME_TRAVEL = 'travel_places';
+export const GROUP_NAME_ACTIVITIES = 'activities';
+export const GROUP_NAME_OBJECTS = 'objects';
+export const GROUP_NAME_SYMBOLS = 'symbols';
+export const GROUP_NAME_FLAGS = 'flags';
+export const GROUP_NAME_RECENTLY_USED = 'recently_used';
+
+export const PROPERTY_DATA_NAME = 'data-name';
+
+export const GROUP_NAMES_ENGLISH = {
+  [GROUP_NAME_PEOPLE]: 'smileys & people',
+  [GROUP_NAME_NATURE]: 'animals & nature',
+  [GROUP_NAME_FOOD]: 'food & drink',
+  [GROUP_NAME_TRAVEL]: 'travel & places',
+  [GROUP_NAME_OBJECTS]: GROUP_NAME_OBJECTS,
+  [GROUP_NAME_ACTIVITIES]: GROUP_NAME_ACTIVITIES,
+  [GROUP_NAME_SYMBOLS]: GROUP_NAME_SYMBOLS,
+  [GROUP_NAME_FLAGS]: GROUP_NAME_FLAGS,
+  [GROUP_NAME_RECENTLY_USED]: 'Recently Used',
+} as const;
